MapOverlayBar from index.js to MapOverlayBar.js
